+++ Source/Sitecore.FakeDb.Tests/GettingStarted.cs
@@ -130,22 +130,6 @@
       }
     }
 
-    [Fact]
-    public void HowDoIMockAuthenticationProvider()
-    {
-      // create mock and configure behaviour of the authentication provider
-      var provider = Substitute.For<Sitecore.Security.Authentication.AuthenticationProvider>();
-      provider.Login("John", false).Returns(true);
-
-      // substitute authentication provider with mock
-      using (new Sitecore.Common.Switcher<Sitecore.Security.Authentication.AuthenticationProvider>(provider))
-      {
-        // use authentication manager in your code
-        bool isLoggedIn = Sitecore.Security.Authentication.AuthenticationManager.Login("John", false);
-        Assert.True(isLoggedIn);
-      }
-    }
-
     #endregion
 
     #region Pipelines